- corrected sql errors and php warnings
git-svn-id: https://svn.typo3.org/TYPO3v4/Extensions/meta_feedit/trunk@88447 735d13b6-9817-0410-8766-e36946ffe9aa

+++ class.tx_metafeedit_export.php
@@ -529,7 +529,7 @@ class tx_metafeedit_pdf extends FPDF {
 	{
 		parent::Close();
 		// clean up tmp files
-		foreach($this->tmpFiles as $tmp) @unlink($tmp);
+		if (is_array($this->tmpFiles)) foreach($this->tmpFiles as $tmp) @unlink($tmp);
 	}
 	
 	/*******************************************************************************
